cleaned up assets and imports
