+++ package.json
@@ -9,10 +9,9 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "bulma": "^0.9.4",
-    "normalize.css": "^8.0.1",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-query": "^3.39.3"
   },
   "devDependencies": {
     "@types/react": "^18.0.24",
