--- src/components/LooksRare.tsx
@@ -1,11 +1,0 @@
-const LooksRare = () => {
-  return (
-    <svg viewBox="0 0 96 96" fill="none" xmlns="http://www.w3.org/2000/svg" className="fill-slate-800 hover:fill-primary">
-      <path fill-rule="evenodd" clip-rule="evenodd" d="M48 55.7837C39.7639 55.7837 33.0811 49.1076 33.0811 40.8648C33.0811 32.6219 39.7639 25.9458 48 25.9458C56.2361 25.9458 62.9189 32.6219 62.9189 40.8648C62.9189 49.1076 56.2361 55.7837 48 55.7837ZM41.5135 40.8648C41.5135 44.4486 44.4191 47.3512 48 47.3512C51.5809 47.3512 54.4865 44.4486 54.4865 40.8648C54.4865 37.2809 51.5809 34.3783 48 34.3783C44.4191 34.3783 41.5135 37.2809 41.5135 40.8648Z" />
-      <path fill-rule="evenodd" clip-rule="evenodd" d="M0 40.8845L28.5405 12.3242H67.4595L96 40.8845L48 88.8648L0 40.8845ZM70.0541 30.4862C57.9274 18.3063 38.0726 18.3064 25.9459 30.4863L15.5676 40.8648L25.9459 51.2433C38.0726 63.4232 57.9274 63.4231 70.0541 51.2432L80.4324 40.8648L70.0541 30.4862Z" />
-    </svg>
-
-  )
-}
-
-export default LooksRare
